proto break 0060 ( 2.99.2) send times for lag comp as doubles cus doubles ROCK.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@16956 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/SyncClock.cxx
@@ -44,7 +44,7 @@ void ServerSyncedClock::update ( ServerLink *link )
 	}
 }
 
-void ServerSyncedClock::timeMessage ( unsigned char tag, float time )
+void ServerSyncedClock::timeMessage ( unsigned char tag, double time )
 {
 	double sentTime = -1;
 	double current = TimeKeeper::getCurrent().getSeconds();
